feat!: do not close WaitChan before Ready
This commit changes when WaitChan is allowed to close. It now requires
Ready instead of a single Add/Done.

It will be a breaking change and released as v2.

This prevents a race while adding tasks. e.g.:

	wg.Go(func() {
	})
	wg.Go(func() {
		// This Add(1) might be executed after the previous task
		// already completed, resulting in WaitChan being closed
		// and a panic.
	})
	<-wg.WaitChan

The only way to prevent against this in v1 is to wrap the task adding
with wg.Add(1) and wg.Done() which seems strange/superfluous. E.g.

	wg.Add(1)
	wg.Go(func() {
	})
	wg.Go(func() {
	})
	wg.Done()
	<-wg.WaitChan

Instead, we now do the following:

	wg.Go(func() {
	})
	wg.Go(func() {
	})
	wg.Ready()
	<-wg.WaitChan

It was too easy to shoot yourself in the foot with WaitChan closing
before all tasks were added.

BREAKING CHANGE: A call to Ready() is now required to make WaitChan
complete.